requirements.txt
openai>=1.0.0
fal-client>=0.4.0
requests>=2.28.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
Pillow>=10.0.0
python-multipart>=0.0.9
